Changed - Abilities now make reactions
Multiple debuffs can be applied to every enemy.
Certain combinations of debuffs cause different reactions.
Added electric tower.
Added tower animations.

+++ Assets/2. Scripts/Interfaces/IHasHealth.cs
@@ -2,8 +2,8 @@ using System;
 
 public interface IHasHealth
 {
-    int Current { get; }
+    float Current { get; }
     int Max { get; }
-    event Action<int, int> OnHealthChanged;
+    event Action<float, int> OnHealthChanged;
     event Action OnDeath;
 }
