+++ Assets/2. Scripts/Systems/Enemies/Health.cs
@@ -3,9 +3,9 @@ using System;
 public class Health : IHasHealth
 {
     public int Max { get;  }
-    public int Current { get; private set; }
+    public float Current { get; private set; }
 
-    public event Action<int, int> OnHealthChanged; //Current, max
+    public event Action<float, int> OnHealthChanged; //Current, max
     public event Action OnDeath;
 
     public bool IsAlive => Current > 0;
@@ -16,7 +16,7 @@ public class Health : IHasHealth
         Current = Max;
     }
 
-    public void TakeDamage(int amount)
+    public void TakeDamage(float amount)
     {
         if (amount <= 0 || Current <= 0) return;
         Current = Math.Max(0, Current - amount);
